Create TBS through constructor instead of init_...
