+++ openfisca_france/scripts/getting_started/test1.py
@@ -2,7 +2,7 @@
 
 import openfisca_france
 
-tax_benefit_system = openfisca_france.init_tax_benefit_system()
+tax_benefit_system = openfisca_france.FranceTaxBenefitSystem()
 
 scenario = tax_benefit_system.new_scenario()
 scenario.init_single_entity(
