+++ openfisca_france/scripts/link_ipp_tax_and_benefit_tables_to_parameters.py
@@ -15,7 +15,7 @@ import sys
 import yaml
 
 
-from openfisca_france import init_tax_benefit_system
+from openfisca_france import FranceTaxBenefitSystem
 
 
 date_names = (
@@ -181,7 +181,7 @@ def main():
 
 #    print yaml.dump(ipp_infos_by_value, allow_unicode = True, default_flow_style = False, indent = 2, width = 120)
 
-    tax_benefit_system = init_tax_benefit_system()
+    tax_benefit_system = FranceTaxBenefitSystem()
 
     # print yaml.dump(tax_benefit_system.legislation_json, allow_unicode = True, default_flow_style = False, indent = 2,
     #     width = 120)
